Add support to get command to pass multi-valued parameters as the query string to an HTTP GET.
Query parameters can be multi-valued, and this commit includes tests to
validate functionality of passing a single valued parameter, as well as
a multivalued parameter.

Multi-valued parameters can be passed in the form of a JSON array in the
"params" option to the "get" command.

As an example:
get FooContext --params "{test: ['value1', 'value2']}"